temp

+++ main.go
@@ -215,8 +215,8 @@ func Randint(min, max int) int {
 
 // Get a random integer between min and max for int64
 //
-// example usage: temp := ez.Randint64(1, 10)
-func Randint64(min, max int) int64 {
+// example usage: temp := ez.Randint64(int64(1), int64(10))
+func Randint64(min, max int64) int64 {
 	return min + int64(fastrand.Uint64n(uint64(max-min+1)))
 }
